requirements.txt
cython
cryptofeed
order-book
requests
pyyaml
fastapi
pytest
uvicorn[standard]
numpy
pyzmq
numba
